refactor: реорганизация пакетов и улучшение архитектуры экрана фильтрации
Изменения в структуре пакетов:
- Перемещение FiltersViewModel в пакет presentation.viewmodel
- Выделение FiltersActions и FiltersUiState в отдельные файлы внутри presentation.filters
- Выделение IndustryActions в отдельный файл

Изменения в FiltersEntryProvider:
- Обновление FiltersActions: onWorkLocationFilter → onWorkLocationClick, onIndustryFilter → onIndustryClick, onCheckBox → onCheckBoxChange
- Удаление неиспользуемого onSearchTextChange
- Передача колбэков через IndustryActions в SelectIndustryScreen

Изменения в SelectIndustryScreen:
- Вынос внутреннего компонента ShowContent в отдельный файл IndustrySelectionList
- Упрощение сигнатуры: теперь принимает IndustryActions вместо отдельных колбэков
- Вынос логики списка отраслей в переиспользуемый компонент
