+++ app/src/main/java/ru/practicum/android/diploma/feature/filters/ui/industry/SelectIndustryScreen.kt
@@ -2,27 +2,21 @@ package ru.practicum.android.diploma.feature.filters.ui.industry
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.stringResource
 import ru.practicum.android.diploma.R
 import ru.practicum.android.diploma.app.ui.theme.AppDimensions
 import ru.practicum.android.diploma.core.domain.model.FilterIndustry
 import ru.practicum.android.diploma.core.presentation.components.AppSearchBar
 import ru.practicum.android.diploma.core.presentation.components.AppTopBar
+import ru.practicum.android.diploma.feature.filters.presentation.industry.IndustryActions
 import ru.practicum.android.diploma.feature.filters.presentation.industry.IndustryScreenState
 import ru.practicum.android.diploma.feature.filters.presentation.industry.IndustryUiState
 import ru.practicum.android.diploma.feature.filters.ui.ApplyButton
-import ru.practicum.android.diploma.feature.filters.ui.filters.RadioButtonItem
 import ru.practicum.android.diploma.feature.filters.ui.states.FilterFetchErrorState
 
 @Composable
@@ -30,10 +24,7 @@ fun SelectIndustryScreen(
     modifier: Modifier = Modifier,
     screenState: IndustryScreenState,
     initSelectedIndustry: FilterIndustry?,
-    onIndustryClick: (FilterIndustry) -> Unit,
-    onSearchTextChanged: (String) -> Unit,
-    onApplyClick: (FilterIndustry?) -> Unit,
-    onBackClick: () -> Unit
+    actions: IndustryActions
 ) {
     val isApplyButtonEnabled = initSelectedIndustry != screenState.selectedIndustry
 
@@ -41,7 +32,7 @@ fun SelectIndustryScreen(
         topBar = {
             AppTopBar(
                 title = stringResource(R.string.screen_select_industry),
-                onNavigationIcon = onBackClick
+                onNavigationIcon = actions.onBackClick
             )
         }
     ) { paddingValues ->
@@ -51,7 +42,7 @@ fun SelectIndustryScreen(
             AppSearchBar(
                 text = screenState.searchText,
                 hint = stringResource(R.string.hint_search_industry),
-                onTextChange = { onSearchTextChanged(it) }
+                onTextChange = { actions.onSearchTextChanged(it) }
             )
 
             when (val stateContent = screenState.uiState) {
@@ -61,10 +52,10 @@ fun SelectIndustryScreen(
                             .fillMaxWidth()
                             .weight(1f)
                     ) {
-                        ShowContent(
+                        IndustrySelectionList(
                             industries = stateContent.filteredIndustries,
                             selectedIndustry = selectedIndustry,
-                            onSelectionChange = onIndustryClick
+                            onSelectionChange = actions.onIndustryClick
                         )
                     }
 
@@ -72,49 +63,13 @@ fun SelectIndustryScreen(
                         ApplyButton(
                             modifier = Modifier.padding(bottom = AppDimensions.paddingBig),
                             text = stringResource(R.string.button_choose),
-                            onClick = { onApplyClick(selectedIndustry) }
+                            onClick = { actions.onApplyClick(selectedIndustry) }
                         )
                     }
                 }
 
                 IndustryUiState.FetchError -> FilterFetchErrorState()
                 IndustryUiState.Loading -> {}
-            }
-        }
-    }
-}
-
-@Composable
-private fun ShowContent(
-    industries: List<FilterIndustry>,
-    selectedIndustry: FilterIndustry?,
-    onSelectionChange: (FilterIndustry) -> Unit
-) {
-    val keyboardController = LocalSoftwareKeyboardController.current
-    val listState = rememberLazyListState()
-
-    LaunchedEffect(listState.isScrollInProgress) {
-        if (listState.isScrollInProgress) {
-            keyboardController?.hide()
-        }
-    }
-
-    LazyColumn(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(top = AppDimensions.paddingSmall),
-        state = listState
-    ) {
-        items(
-            items = industries,
-            key = { it.id }
-        ) { industry ->
-            RadioButtonItem(
-                text = industry.name,
-                isSelected = industry == selectedIndustry
-            ) {
-                onSelectionChange(industry)
-                keyboardController?.hide()
             }
         }
     }
